fix: vérification de l'existence d'un fichier avant suppression par shred (#2236)

+++ .bin/scripts/merge-vault.sh
@@ -20,7 +20,9 @@ chmod 600 $pass_file
 
 delete_tempfiles() {
   rm -f "$ancestor_tempfile" "$current_tempfile" "$other_tempfile"
-  shred -f -n 10 -u "$pass_file"
+  if [ -f "$pass_file" ]; then
+    shred -f -n 10 -u "$pass_file"
+  fi
 }
 trap delete_tempfiles EXIT
